Routine development.

+++ db/PostgreSQL/2.x/2.0/src/02.functions-and-logic/finance.get_account_id_by_account_name.sql
@@ -1,7 +1,7 @@
 ﻿DROP FUNCTION IF EXISTS finance.get_account_id_by_account_name(text);
 
 CREATE FUNCTION finance.get_account_id_by_account_name(text)
-RETURNS bigint
+RETURNS integer
 STABLE
 AS
 $$
