+++ db/PostgreSQL/2.x/2.0/src/02.functions-and-logic/finance.get_account_id_by_account_number.sql
@@ -1,7 +1,7 @@
 ﻿DROP FUNCTION IF EXISTS finance.get_account_id_by_account_number(text);
 
 CREATE FUNCTION finance.get_account_id_by_account_number(text)
-RETURNS bigint
+RETURNS integer
 STABLE
 AS
 $$
